Remove OwnerId from Plant model, add Create action for categories and implement [Authorize] for both

+++ UrbanPlantRescueApp/Controllers/PlantController.cs
@@ -1,23 +1,31 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+using UrbanPlantRescueApp.Data;
 using UrbanPlantRescueApp.Services;
 using UrbanPlantRescueApp.Services.Interfaces;
 using UrbanPlantRescueApp.ViewModels;
 
 namespace UrbanPlantRescueApp.Controllers
 {
+    [Authorize]
     public class PlantController : Controller
     {
+        private readonly ApplicationDbContext dbContext;
         private readonly IPlantService plantService;
         private readonly ICategoryService categoryService;
-        public PlantController(IPlantService plantService, ICategoryService categoryService)
+        public PlantController(ApplicationDbContext dbContext, IPlantService plantService, ICategoryService categoryService)
         {
+            this.dbContext = dbContext;
             this.plantService = plantService;
             this.categoryService = categoryService;
         }
+        [AllowAnonymous]
         public async Task<IActionResult> Index()
         {
-            var plants = await plantService.GetAllPlantsAsync();
-            return View(plants);
+            var allPlants = await plantService.GetAllPlantsAsync();
+            return View(allPlants);
         }
         public async Task<IActionResult> Details(int id)
         {
@@ -60,6 +68,8 @@ namespace UrbanPlantRescueApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(int id, PlantFormViewModel model)
         {
+            var plant = await dbContext.Plants.FindAsync(id);
+            if (plant == null) { return NotFound(); }
             if (!ModelState.IsValid)
             {
                 model.Categories = await categoryService.GetAllCategoriesAsync();
@@ -79,6 +89,11 @@ namespace UrbanPlantRescueApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var plant = await dbContext.Plants.FindAsync(id);
+            if (plant == null)
+            {
+                return NotFound();
+            }
             await plantService.DeletePlantAsync(id);
             return RedirectToAction(nameof(Index));
         }
